Added notes.
