+++ TravelGroup/TravelGroup/LogDetail/LogDetailViewController.swift
@@ -13,18 +13,21 @@ class LogDetailViewController: UIViewController {
     @IBOutlet weak var logTitleTextField: UITextField!
     @IBOutlet weak var logAddressTextField: UITextField!
     @IBOutlet weak var logBodyTextView: UITextView!
-    
+    // Pulling over the outlets for the detail screen.
     
     // MARK: - Lifecycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        // calling updateUI
         updateUI()
     }
     // MARK: - Properties
     var log: Log?
+    // creating a property to use when we need from our Model.
     
     // MARK: - Actions
+    // This action is saving our title, address and body once it is created.
     @IBAction func saveButtonTapped(_ sender: Any) {
         guard let title = logTitleTextField.text, !title.isEmpty,
               let address = logAddressTextField.text, !address.isEmpty,
@@ -37,15 +40,17 @@ class LogDetailViewController: UIViewController {
             // log on line 26 does NOT have a value, which means the user has tapped on the plus button and is creating a new Diary log!
             LogController.shared.createLog(title: title, address: address, body: body)
         }
+        // Will pop the view like a stack of pancakes.
         navigationController?.popViewController(animated: true)
     }// SaveButtonTapped
     
     // MARK: - Functions
+    // Creating a function updateUI
     func updateUI() {
-        logDateLabel.text = Date().asString()
+        logDateLabel.text = Date().asString() // using our date helper to create a string instead of an INT
         guard let log = log else { return }
         logDateLabel.text = log.logDate.asString()
-        logTitleTextField.text = log.logTitle
+        logTitleTextField.text = log.logTitle // taking the label text field.text and giving it a log. logTitle
         logAddressTextField.text = log.logAddress
         logBodyTextView.text = log.logBody
